moved handlers into closerelease component and fetched collection outside of collector modal to share data with releasePurchase component
