+++ js/web/components/CollectorModal.js
@@ -1,49 +1,17 @@
-import React, { useState, useContext, useEffect } from 'react'
+import React, { useState, useContext } from 'react'
 import { styled } from '@mui/material/styles'
 import Modal from '@mui/material/Modal'
 import Backdrop from '@mui/material/Backdrop'
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import Paper from '@mui/material/Paper'
-import { useWallet } from '@solana/wallet-adapter-react'
 import Nina from '@nina-protocol/nina-internal-sdk/esm/Nina'
-import Release from '@nina-protocol/nina-internal-sdk/esm/Release'
 import Link from 'next/link'
-import { Button } from '@mui/material'
 const CollectorModal = (props) => {
-  const { releasePubkey, metadata } = props
-  const wallet = useWallet()
-  const { collection, displayNameForAccount } = useContext(Nina.Context)
-  const { getCollectorsForRelease } = useContext(Release.Context)
-  const [open, setOpen] = useState(false)
-  const [collectors, setCollectors] = useState()
-  useEffect(() => {
-    handleGetCollectorsForRelease(releasePubkey)
-  }, [collection])
+  const { metadata, collectors } = props
+  const { displayNameForAccount } = useContext(Nina.Context)
 
-  const handleGetCollectorsForRelease = async (releasePubkey) => {
-    const collectorsList = await getCollectorsForRelease(releasePubkey)
-    // Manually check if in user collection since script only updates collectors every hour
-    // and websocket to update on each purchase can be inconsistent
-    if (wallet?.publicKey) {
-      const walletPublicKey = wallet.publicKey.toBase58()
-      if (
-        collection[releasePubkey] > 0 &&
-        !collectorsList.includes(walletPublicKey)
-      ) {
-        collectorsList.push(walletPublicKey)
-      } else if (
-        collectorsList.includes(walletPublicKey) &&
-        collection[releasePubkey] <= 0
-      ) {
-        const index = collectorsList.indexOf(walletPublicKey)
-        if (index > -1) {
-          collectorsList.splice(index, 1)
-        }
-      }
-    }
-    setCollectors(collectorsList)
-  }
+  const [open, setOpen] = useState(false)
   return (
     <>
       {collectors?.length > 0 ? (
